requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.12
django-allauth==0.49.0
django-appconf==1.0.5
django-compressor==3.1
django-countries-plus==1.3.2
django-debug-toolbar==3.2.4
django-elasticsearch-debug-toolbar==3.0.2
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
django-extensions==3.1.5
django-languages-plus==1.1.1
django-sass-processor==1.1
djangorestframework==3.13.1
elastic-apm==6.8.0
elasticsearch==7.17.1
elasticsearch-dsl==7.4.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
libsass==0.21.0
oauthlib==3.2.0
Pillow==9.0.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
rcssmin==1.1.0
requests==2.27.1
requests-oauthlib==1.3.1
rjsmin==1.2.0
sentry-sdk==1.5.7
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==6.0.0
zope.event==4.5.0
zope.interface==5.4.0
